feature: move mixin to support for version
